+++ scripts/run-make.sh
@@ -1,0 +1,9 @@
+#!/bin/sh
+set -eu
+
+exec /usr/bin/env \
+  -u MAKEFILES \
+  -u MAKEFLAGS \
+  -u MFLAGS \
+  -u MAKEOVERRIDES \
+  /usr/bin/make "$@"